third_party/blink/web_tests/fast/css/getComputedStyle/computed-style-select-overflow-expected.txt

Test computed style for the overflow property of select element

PASS computedOverflowStyle('scroll', 'overflow-x') is 'scroll'
PASS computedOverflowStyle('scroll', 'overflow-y') is 'scroll'
PASS computedInheritedOverflowStyle('inherit', 'overflow-x') is 'scroll'
PASS computedInheritedOverflowStyle('inherit', 'overflow-y') is 'scroll'
PASS successfullyParsed is true

TEST COMPLETE

One
Two
Three
Four
Five
Six
Seven
Eight
Nine
Ten
